Use `assert()` to check whether a target node is defined
This makes the editor pause if the condition isn't met, instead of
just printing an error message in the output log. Since defining
a target node is absolutely required for the add-on to work, it
makes sense to pause the editor in this case.

A message can still be displayed thanks to Godot 3.2's custom assert
messages feature.