Fix for frame rate below physics tick rate
There was a bug in the dirty flag logic which could result in a 'double refresh' where frame rate fell below physics tick rate - effectively turning off smoothing in that situation.

This fix removes the dirty flag and simply refreshes on each physics tick. This will do as a sticky plaster as this should be in Godot core very soon anyway.